Upgrade to Ruby version 2.7.4 (#550)
* Remove calls to number_to_human_size

Due to the upgrade of Ruby and the missing upgrade of Rails,
number_to_human_size raises an exception due to a missing
method in the upgraded BigDecimal version. This should be
changed back when Rails is upgraded to >= 5.1.0.

* Add call to presence to return nil instead of empty objects.

Due to the Ruby upgrade the URL when uploading files was broken.

* Change URL for deploy to newstagingsapphire

* Add Ruby version 2.7.0 to be tested

* Change setup-ruby action

actions/setup-ruby is deprecated and should be changed
to ruby/setup-ruby as this is being actively maintained.

* Adding eager loading

In production the lib folder is not getting
loaded automatically.

* Remove calls to number_to_human

Due to the upgrade of Ruby and the missing upgrade of Rails,
number_to_human raises an exception due to a missing
method in the upgraded BigDecimal version. This should be
changed back when Rails is upgraded to >= 5.1.0.

* Upgrade workflow to use Ruby 2.7.4